+++ frontend/src/components/AllRestaurantsPage/AllRestaurantsPage.css
@@ -1,0 +1,17 @@
+.page-container {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  box-sizing: border-box;
+}
+.allRestaurants-left-section {
+  width: calc(2 / 3 * 100%);
+}
+.allRestaurants-right-section {
+  width: calc(1 / 3 * 100%);
+}
+.restaurant-cards-container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2%;
+}
