+++ frontend/src/components/AllRestaurantsPage/Ratings/Ratings.css
@@ -3,9 +3,11 @@
   object-position: top;
   object-fit: cover;
   height: 22px;
+  padding: 3px 3px 3px 10px;
 }
 .rating-icon-img-container {
-  width: 100px;
-  height: 22px;
+  /* width: 100px; */
+  height: 25px;
   overflow: hidden;
+  box-sizing: border-box;
 }
