+++ frontend/src/components/AllRestaurantsPage/index.js
@@ -4,8 +4,9 @@ import { getAllRestaurants } from "../../store/restaurants";
 
 import FilterTabs from "./FilterTabs";
 import Map from "./Map";
-import RestaurantCard from "./RestaurantCard";
+import RestaurantCard from "./RestaurantCard/RestaurantCard";
 import PageNumbers from "./PageNumbers";
+import "./AllRestaurantsPage.css";
 
 const AllRestaurantsPage = () => {
   const dispatch = useDispatch();
@@ -17,13 +18,14 @@ const AllRestaurantsPage = () => {
     (state) => state.restaurants.totalRestaurants
   );
   const [currentPage, setCurrentPage] = useState(1);
-  const [startNum, setStartNum] = useState(1);
+
   const pageNumber = Math.ceil(totalResults / 20); //max 480/20=24
 
   useEffect(() => {
+    // fetch initial page
     dispatch(getAllRestaurants(1));
   }, [dispatch]);
-  console.log(currentPage);
+  // fetch page with offset page
   const handlePageChange = (pageNum) => {
     setCurrentPage(pageNum);
     dispatch(getAllRestaurants(pageNum));
@@ -34,23 +36,27 @@ const AllRestaurantsPage = () => {
   if (!allRestaurants || loadingRestaurants) return <h1>Loading...</h1>;
 
   return (
-    <div>
-      <FilterTabs />
-      {Object.values(allRestaurants).map((restaurant, idx) => (
-        <RestaurantCard
-          key={idx}
-          restaurant={restaurant}
-          idx={displayRestaurantIdx(idx)}
+    <div className="page-container">
+      <div className="allRestaurants-left-section">
+        <FilterTabs />
+        <div className="restaurant-cards-container">
+          {Object.values(allRestaurants).map((restaurant, idx) => (
+            <RestaurantCard
+              key={idx}
+              restaurant={restaurant}
+              idx={displayRestaurantIdx(idx)}
+            />
+          ))}
+        </div>
+        <PageNumbers
+          pageNumber={pageNumber}
+          onPageChange={handlePageChange}
+          currentPage={currentPage}
         />
-      ))}
-      <PageNumbers
-        pageNumber={pageNumber}
-        onPageChange={handlePageChange}
-        currentPage={currentPage}
-        startNum={startNum}
-        setStartNum={setStartNum}
-      />
-      <Map />
+      </div>
+      <div className="allRestaurants-right-section">
+        <Map />
+      </div>
     </div>
   );
 };
